rename bids reader
