+++ python/loris_bids_importer/pyproject.toml
@@ -6,7 +6,7 @@ readme = "README.md"
 license = "GPL-3.0-or-later"
 requires-python = ">= 3.11"
 dependencies = [
-    "loris-bids-reader",
+    "loris-bids-utils",
     "loris-utils",
 ]
