+++ python/loris_bids_importer/src/loris_bids_importer/events.py
@@ -19,8 +19,8 @@ from lib.physio.events import (
 )
 from lib.physio.hed import TagGroupMember, build_hed_tag_groups, filter_inherited_tags
 from lib.physio.parameters import insert_physio_file_parameter, insert_physio_project_parameter
-from loris_bids_reader.files.events import OPTIONAL_EVENT_FIELDS, BidsEventsTsvFile
-from loris_bids_reader.json import BidsJsonFile
+from loris_bids_utils.files.events import OPTIONAL_EVENT_FIELDS, BidsEventsTsvFile
+from loris_bids_utils.json import BidsJsonFile
 from loris_utils.crypto import compute_file_blake2b_hash
 
 from loris_bids_importer.copy_files import copy_loris_bids_file, get_loris_bids_root_file_path
