+++ python/loris_bids_importer/src/loris_bids_importer/mri/sidecar.py
@@ -5,7 +5,7 @@ from lib.env import Env
 from lib.get_session_info import SessionInfo, get_session_info
 from lib.imaging_lib.file_parameter import map_bids_to_loris_file_parameters
 from lib.imaging_lib.mri_scanner import MriScannerInfo
-from loris_bids_reader.mri.sidecar import BidsMriSidecarJsonFile
+from loris_bids_utils.mri.sidecar import BidsMriSidecarJsonFile
 
 
 def get_bids_mri_sidecar_scanner_info(sidecar: BidsMriSidecarJsonFile) -> MriScannerInfo:
