--- python/loris_bids_reader/README.md
@@ -1,3 +1,0 @@
-# LORIS BIDS reader
-
-This package contains a LORIS-agnostic BIDS reader library used to read BIDS datasets and files in order to use these datasets in various LORIS pipelines.
