+++ python/loris_bids_utils/README.md
@@ -1,0 +1,3 @@
+# LORIS BIDS utilities
+
+The LORIS BIDS utilities package provides various BIDS-related utilities such as the ability to read and traverse BIDS datasets or parse BIDS files, which can then be used in various LORIS pipelines.
