+++ python/loris_bids_utils/pyproject.toml
@@ -1,7 +1,7 @@
 [project]
-name = "loris-bids-reader"
+name = "loris-bids-utils"
 version = "27.0.0"
-description = "The LORIS BIDS reader library"
+description = "The LORIS BIDS utilities library"
 readme = "README.md"
 license = "GPL-3.0-or-later"
 requires-python = ">= 3.11"
@@ -18,7 +18,7 @@ build-backend = "hatchling.build"
 allow-direct-references = true
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/loris_bids_reader"]
+packages = ["src/loris_bids_utils"]
 
 [tool.ruff]
 extend = "../../pyproject.toml"
