+++ python/loris_bids_utils/src/loris_bids_utils/files/dataset_description.py
@@ -1,6 +1,6 @@
 from typing import Any
 
-from loris_bids_reader.json import BidsJsonFile
+from loris_bids_utils.json import BidsJsonFile
 
 
 class BidsDatasetDescriptionJsonFile(BidsJsonFile):
