+++ python/loris_bids_utils/src/loris_bids_utils/files/participants.py
@@ -5,7 +5,7 @@ import dateutil.parser
 from dateutil.parser import ParserError
 from loris_utils.iter import find, replace_or_append
 
-from loris_bids_reader.tsv import BidsTsvFile, BidsTsvRow
+from loris_bids_utils.tsv import BidsTsvFile, BidsTsvRow
 
 
 class BidsParticipantTsvRow(BidsTsvRow):
