+++ python/loris_bids_utils/src/loris_bids_utils/utils.py
